feat: Finalize moving sections to their own components for landing page and structure LandingPage component folder better
